manifest/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
dnspython==2.3.0
gevent==22.10.2
greenlet==2.0.2
grequests==0.6.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
packaging==23.1
prompt-toolkit==3.0.38
pymongo==4.3.3
pytz==2023.3
redis==4.3.4
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zope.event==4.6
zope.interface==6.0
